// frontend/src/proto/gen/google/rpc/error_details_pb.ts
export type BadRequest = Message<'google.rpc.BadRequest'> & {
    /**
     * Describes all violations in a client request.
     *
     * @generated from field: repeated google.rpc.BadRequest.FieldViolation field_violations = 1;
     */
    fieldViolations: BadRequest_FieldViolation[];
};

/**
 * Describes the message google.rpc.BadRequest.
 * Use `create(BadRequestSchema)` to create a new message.
 */
export const BadRequestSchema: GenMessage<BadRequest> = /*@__PURE__*/ messageDesc(file_google_rpc_error_details, 5);

/**
 * A message type used to describe a single bad request field.
 *
 * @generated from message google.rpc.BadRequest.FieldViolation
 */
export type BadRequest_FieldViolation = Message<'google.rpc.BadRequest.FieldViolation'> & {
    /**
     * A path that leads to a field in the request body. The value will be a
     * sequence of dot-separated identifiers that identify a protocol buffer
     * field.
     *
     * Consider the following:
     *
     *     message CreateContactRequest {
     *       message EmailAddress {
     *         enum Type {
     *           TYPE_UNSPECIFIED = 0;
     *           HOME = 1;
     *           WORK = 2;
     *         }
     *
     *         optional string email = 1;
     *         repeated EmailType type = 2;
     *       }
     *
     *       string full_name = 1;
     *       repeated EmailAddress email_addresses = 2;
     *     }
     *
     * In this example, in proto `field` could take one of the following values:
     *
     * * `full_name` for a violation in the `full_name` value
     * * `email_addresses[1].email` for a violation in the `email` field of the
     *   first `email_addresses` message
     * * `email_addresses[3].type[2]` for a violation in the second `type`
     *   value in the third `email_addresses` message.
     *
     * In JSON, the same values are represented as:
     *
     * * `fullName` for a violation in the `fullName` value
     * * `emailAddresses[1].email` for a violation in the `email` field of the
     *   first `emailAddresses` message
     * * `emailAddresses[3].type[2]` for a violation in the second `type`
     *   value in the third `emailAddresses` message.
     *
     * @generated from field: string field = 1;
     */
    field: string;

    /**
     * A description of why the request element is bad.
     *
     * @generated from field: string description = 2;
     */
    description: string;

    /**
     * The reason of the field-level error. This is a constant value that
     * identifies the proximate cause of the field-level error. It should
     * uniquely identify the type of the FieldViolation within the scope of the
     * google.rpc.ErrorInfo.domain. This should be at most 63
     * characters and match a regular expression of `[A-Z][A-Z0-9_]+[A-Z0-9]`,
     * which represents UPPER_SNAKE_CASE.
     *
     * @generated from field: string reason = 3;
     */
    reason: string;

    /**
     * Provides a localized error message for field-level errors that is safe to
     * return to the API consumer.
     *
     * @generated from field: google.rpc.LocalizedMessage localized_message = 4;
     */
    localizedMessage?: LocalizedMessage;
};
// ...
export type LocalizedMessage = Message<'google.rpc.LocalizedMessage'> & {
    /**
     * The locale used following the specification defined at
     * https://www.rfc-editor.org/rfc/bcp/bcp47.txt.
     * Examples are: "en-US", "fr-CH", "es-MX"
     *
     * @generated from field: string locale = 1;
     */
    locale: string;

    /**
     * The localized error message in the above locale.
     *
     * @generated from field: string message = 2;
     */
    message: string;
};